add basic API for bootstrap, hide sensitive info
