+++ modules/EventEmitter.js
@@ -3,7 +3,6 @@ const events = require('events');
 
 const Challenge = require('./Challenge');
 const Utilities = require('./Utilities');
-const Graph = require('./Graph');
 const Models = require('../models');
 const ImportUtilities = require('./ImportUtilities');
 const ObjectValidator = require('./validator/object-validator');
@@ -97,26 +96,7 @@ class EventEmitter {
             dcService,
             dvController,
             notifyError,
-            transport,
         } = this.ctx;
-
-        this._on('api-node-info', async (data) => {
-            try {
-                const system = await transport.getNetworkInfo();
-                data.response.status(200);
-                data.response.send({
-                    system,
-                    config,
-                });
-            } catch (err) {
-                logger.error('Failed to get node info');
-                notifyError(err);
-                data.response.status(500);
-                data.response.send({
-                    message: err,
-                });
-            }
-        });
 
         this._on('api-network-query-responses', async (data) => {
             const { query_id } = data;
