+++ modules/service/rest-api-service.js
@@ -1,0 +1,478 @@
+const fs = require('fs');
+const ip = require('ip');
+const restify = require('restify');
+const corsMiddleware = require('restify-cors-middleware');
+
+const Utilities = require('../Utilities');
+const pjson = require('../../package.json');
+const RestAPIValidator = require('../validator/rest-api-validator');
+
+class RestAPIService {
+    constructor(ctx) {
+        this.ctx = ctx;
+        this.config = ctx.config;
+        this.logger = ctx.logger;
+        this.apiUtilities = ctx.apiUtilities;
+    }
+
+    /**
+     * Start RPC server
+     */
+    async startRPC() {
+        const options = {
+            name: 'RPC server',
+            version: pjson.version,
+            formatters: {
+                'application/json': (req, res, body) => {
+                    res.set('content-type', 'application/json; charset=utf-8');
+                    if (!body) {
+                        if (res.getHeader('Content-Length') === undefined && res.contentLength === undefined) {
+                            res.setHeader('Content-Length', 0);
+                        }
+                        return null;
+                    }
+
+                    if (body instanceof Error) {
+                        // snoop for RestError or HttpError, but don't rely on instanceof
+                        if ((body.restCode || body.httpCode) && body.body) {
+                            // eslint-disable-next-line
+                            body = body.body;
+                        } else {
+                            body = {
+                                message: body.message,
+                            };
+                        }
+                    }
+
+                    if (Buffer.isBuffer(body)) {
+                        body = body.toString('base64');
+                    }
+
+                    let ident = 2;
+                    if ('prettify-json' in req.headers) {
+                        if (req.headers['prettify-json'] === 'false') {
+                            ident = 0;
+                        }
+                    }
+                    const data = Utilities.stringify(body, ident);
+
+                    if (res.getHeader('Content-Length') === undefined && res.contentLength === undefined) {
+                        res.setHeader('Content-Length', Buffer.byteLength(data));
+                    }
+                    return data;
+                },
+            },
+        };
+
+        if (this.config.node_rpc_use_ssl) {
+            Object.assign(options, {
+                key: fs.readFileSync(this.config.node_rpc_ssl_key_path),
+                certificate: fs.readFileSync(this.config.node_rpc_ssl_cert_path),
+                rejectUnauthorized: true,
+            });
+        }
+
+        const server = restify.createServer(options);
+
+        server.use(restify.plugins.acceptParser(server.acceptable));
+        server.use(restify.plugins.queryParser());
+        server.use(restify.plugins.bodyParser());
+        const cors = corsMiddleware({
+            preflightMaxAge: 5, // Optional
+            origins: ['*'],
+            allowHeaders: ['API-Token', 'prettify-json', 'raw-data'],
+            exposeHeaders: ['API-Token-Expiry'],
+        });
+
+        server.pre(cors.preflight);
+        server.use(cors.actual);
+        server.use((request, response, next) => {
+            const result = this.apiUtilities.authorize(request);
+            if (result) {
+                response.status(result.status);
+                response.send({
+                    message: result.message,
+                });
+                return;
+            }
+            return next();
+        });
+
+        // TODO: Temp solution to listen all adapters in local net.
+        let serverListenAddress = this.config.node_rpc_ip;
+        if (ip.isLoopback(serverListenAddress)) {
+            serverListenAddress = '0.0.0.0';
+        }
+
+        // promisified server.listen()
+        const startServer = () => new Promise((resolve, reject) => {
+            server.listen(this.config.node_rpc_port, serverListenAddress, (err) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    resolve();
+                }
+            });
+        });
+
+        await startServer(server, serverListenAddress);
+        this.logger.notify(`API exposed at  ${server.url}`);
+
+        if (this.config.is_bootstrap_node) {
+            this._exposeBootstrapAPIRoutes(server);
+        } else {
+            this._exposeAPIRoutes(server);
+        }
+    }
+
+    /**
+     * Expose bootstrap API Routes
+     */
+    _exposeBootstrapAPIRoutes(server) {
+        this._registerNodeInfoRoute(server, true);
+    }
+
+    /**
+     * API Routes
+     */
+    _exposeAPIRoutes(server) {
+        const {
+            importController, dcController, transport, emitter,
+        } = this.ctx;
+
+        this._registerNodeInfoRoute(server, false);
+
+        /**
+         * Data import route
+         * @param importfile - file or text data
+         * @param importtype - (GS1/WOT)
+         */
+        server.post('/api/import', async (req, res) => {
+            await importController.import(req, res);
+        });
+
+        /**
+         * Create offer route
+         */
+        server.post('/api/replication', async (req, res) => {
+            await dcController.createOffer(req, res);
+        });
+
+        server.get('/api/dump/rt', (req, res) => {
+            this.logger.api('Dumping routing table');
+            const message = transport.dumpContacts();
+
+            res.status(200);
+            res.send({
+                message,
+            });
+        });
+
+        server.get('/api/network/get-contact/:node_id', async (req, res) => {
+            const nodeId = req.params.node_id;
+            this.logger.api(`Get contact node ID ${nodeId}`);
+
+            const result = await transport.getContact(nodeId);
+            const body = {};
+
+            if (result) {
+                Object.assign(body, result);
+            }
+            res.status(200);
+            res.send(body);
+        });
+
+        server.get('/api/network/find/:node_id', async (req, res) => {
+            const nodeId = req.params.node_id;
+            this.logger.api(`Find node ID ${nodeId}`);
+
+            const result = await transport.findNode(nodeId);
+            const body = {};
+
+            if (result) {
+                Object.assign(body, result);
+            }
+            res.status(200);
+            res.send(body);
+        });
+
+        server.get('/api/replication/:replication_id', (req, res) => {
+            this.logger.api('GET: Replication status request received');
+
+            const replicationId = req.params.replication_id;
+            if (replicationId == null) {
+                this.logger.error('Invalid request. You need to provide replication ID');
+                res.status = 400;
+                res.send({
+                    message: 'Replication ID is not provided',
+                });
+            } else {
+                const queryObject = {
+                    replicationId,
+                    response: res,
+                };
+                emitter.emit('api-offer-status', queryObject);
+            }
+        });
+
+        /**
+         * Get trail from database
+         * @param QueryObject - ex. {uid: abc:123}
+         */
+        server.get('/api/trail', (req, res, next) => {
+            this.logger.api('GET: Trail request received.');
+
+            const error = RestAPIValidator.validateNotEmptyQuery(req.query);
+            if (error) {
+                return next(error);
+            }
+            const queryObject = req.query;
+            emitter.emit('api-trail', {
+                query: queryObject,
+                response: res,
+            });
+        });
+
+        /** Get root hash for provided data query
+         * @param Query params: data_set_id
+         */
+        server.get('/api/fingerprint', (req, res) => {
+            this.logger.api('GET: Fingerprint request received.');
+
+            const queryObject = req.query;
+            emitter.emit('api-get_root_hash', {
+                query: queryObject,
+                response: res,
+            });
+        });
+
+        server.get('/api/query/network/:query_id', (req, res) => {
+            this.logger.api('GET: Query for status request received.');
+
+            if (!req.params.query_id) {
+                res.status(400);
+                res.send({
+                    message: 'Param required.',
+                });
+                return;
+            }
+            emitter.emit('api-network-query-status', {
+                id: req.params.query_id,
+                response: res,
+            });
+        });
+
+        server.get('/api/query/:query_id/responses', (req, res) => {
+            this.logger.api('GET: Local query responses request received.');
+
+            if (!req.params.query_id) {
+                res.status(400);
+                res.send({
+                    message: 'Param query_id is required.',
+                });
+                return;
+            }
+            emitter.emit('api-network-query-responses', {
+                query_id: req.params.query_id,
+                response: res,
+            });
+        });
+
+        server.post('/api/query/network', (req, res, next) => {
+            this.logger.api('POST: Network query request received.');
+
+            let error = RestAPIValidator.validateBodyRequired(req.body);
+            if (error) {
+                return next(error);
+            }
+
+            const { query } = req.body;
+            error = RestAPIValidator.validateSearchQuery(query);
+            if (error) {
+                return next(error);
+            }
+
+            emitter.emit('api-network-query', {
+                query,
+                response: res,
+            });
+        });
+
+        /**
+         * Get vertices by query
+         * @param queryObject
+         */
+        server.post('/api/query/local', (req, res, next) => {
+            this.logger.api('POST: Local query request received.');
+
+            let error = RestAPIValidator.validateBodyRequired(req.body);
+            if (error) {
+                return next(error);
+            }
+
+            const queryObject = req.body.query;
+            error = RestAPIValidator.validateSearchQuery(queryObject);
+            if (error) {
+                return next(error);
+            }
+
+            // TODO: Decrypt returned vertices
+            emitter.emit('api-query', {
+                query: queryObject,
+                response: res,
+            });
+        });
+
+        server.get('/api/query/local/import/:data_set_id', (req, res) => {
+            this.logger.api('GET: Local import request received.');
+
+            if (!req.params.data_set_id) {
+                res.status(400);
+                res.send({
+                    message: 'Param required.',
+                });
+                return;
+            }
+
+            emitter.emit('api-query-local-import', {
+                data_set_id: req.params.data_set_id,
+                request: req,
+                response: res,
+            });
+        });
+
+        server.post('/api/read/network', (req, res) => {
+            this.logger.api('POST: Network read request received.');
+
+            if (req.body == null || req.body.query_id == null || req.body.reply_id == null
+                || req.body.data_set_id == null) {
+                res.status(400);
+                res.send({ message: 'Bad request' });
+                return;
+            }
+            const { query_id, reply_id, data_set_id } = req.body;
+
+            emitter.emit('api-choose-offer', {
+                query_id,
+                reply_id,
+                data_set_id,
+                response: res,
+            });
+        });
+
+
+        server.post('/api/deposit', (req, res) => {
+            this.logger.api('POST: Deposit tokens request received.');
+
+            if (req.body !== null && typeof req.body.trac_amount === 'number'
+                && req.body.trac_amount > 0) {
+                const { trac_amount } = req.body;
+                emitter.emit('api-deposit-tokens', {
+                    trac_amount,
+                    response: res,
+                });
+            } else {
+                res.status(400);
+                res.send({ message: 'Bad request' });
+            }
+        });
+
+
+        server.post('/api/withdraw', (req, res) => {
+            this.logger.api('POST: Withdraw tokens request received.');
+
+            if (req.body !== null && typeof req.body.trac_amount === 'number'
+                && req.body.trac_amount > 0) {
+                const { trac_amount } = req.body;
+                emitter.emit('api-withdraw-tokens', {
+                    trac_amount,
+                    response: res,
+                });
+            } else {
+                res.status(400);
+                res.send({ message: 'Bad request' });
+            }
+        });
+
+        server.get('/api/import_info', async (req, res) => {
+            await importController.dataSetInfo(req, res);
+        });
+
+        server.get('/api/imports_info', (req, res) => {
+            this.logger.api('GET: List imports request received.');
+
+            emitter.emit('api-imports-info', {
+                response: res,
+            });
+        });
+
+        server.get('/api/consensus/:sender_id', (req, res) => {
+            this.logger.api('GET: Consensus check events request received.');
+
+            if (req.params.sender_id == null) {
+                res.status(400);
+                res.send({ message: 'Bad request' });
+            }
+
+            emitter.emit('api-consensus-events', {
+                sender_id: req.params.sender_id,
+                response: res,
+            });
+        });
+
+        /**
+         * Temporary route used for HTTP network prototype
+         */
+        server.post('/network/send', (req, res) => {
+            this.logger.api('P2P request received');
+
+            const { type } = req.body;
+            emitter.emit(type, req, res);
+        });
+    }
+
+    /**
+     * Register common info route
+     * @param server - Server instance
+     * @param isBootstrap - Is this a bootstrap node?
+     * @private
+     */
+    _registerNodeInfoRoute(server, isBootstrap) {
+        const {
+            transport,
+            config,
+        } = this.ctx;
+
+        server.get('/api/info', async (req, res) => {
+            this.logger.api('GET: Node information request received.');
+
+            try {
+                const network = await transport.getNetworkInfo();
+                const basicConfig = {
+                    version: pjson.version,
+                    blockchain: config.blockchain.blockchain_title,
+                    network,
+                    is_bootstrap: isBootstrap,
+                };
+
+                if (!isBootstrap) {
+                    Object.assign(basicConfig, {
+                        node_wallet: config.node_wallet,
+                        erc_725_identity: config.erc725Identity,
+                    });
+                }
+
+                res.status(200);
+                res.send(basicConfig);
+            } catch (e) {
+                this.logger.error('Failed to process /api/info route', e);
+                res.status(500);
+                res.send({
+                    message: `Failed to process /api/info route. Error ${e.message}`,
+                });
+            }
+        });
+    }
+}
+
+module.exports = RestAPIService;
